requirements.txt
openai==1.6.1
streamlit-option-menu==0.3.6
st-star-rating==0.0.6